requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
Flask==0.10.1
Flask-Celery-Helper==1.1.0
flask-peewee==0.6.6
itsdangerous==0.24
Jinja2==2.7.3
kombu==3.0.26
MarkupSafe==0.23
peewee==2.6.1
pytz==2015.4
redis==2.10.3
Werkzeug==0.10.4
wheel==0.24.0
wtf-peewee==0.2.3
WTForms==2.0.2
